+++ tests/test_module.py
@@ -36,17 +36,23 @@ def test_input_validation_on_repository(project: Project):
     yum::Repository(name="test")
         """
         )
-    assert "baseurl and mirrorlist cannot be null at the same time." in str(
+    assert "baseurl, mirrorlist and metalink cannot be null at the same time." in str(
         excinfo.value
     )
 
 
-def test_repository(project: Project):
+@pytest.mark.parametrize("baseurl", [True, False])
+@pytest.mark.parametrize("mirrorlist", [True, False])
+@pytest.mark.parametrize("metalink", [True, False])
+def test_repository(project: Project, baseurl: bool, mirrorlist: bool, metalink: bool):
     """
     Basic test for the Repository entity.
     """
+    if not baseurl and not mirrorlist and not metalink:
+        # This combination is tested in test case: test_input_validation_on_repository
+        return
     project.compile(
-        """
+        f"""
 import std
 import yum
 import redhat
@@ -56,8 +62,9 @@ yum::Repository(
     name="test",
     gpgcheck=true,
     enabled=true,
-    baseurl="http://baseurl.com",
-    mirrorlist="http://mirror.com",
+    baseurl={"'http://baseurl.com'" if baseurl else "null"},
+    mirrorlist={"'http://mirror.com'" if mirrorlist else "null"},
+    metalink={"'http://metalink.com'" if metalink else "null"},
     gpgkey="http://gpgkey.com",
     metadata_expire=7200,
     skip_if_unavailable=false,
@@ -65,18 +72,26 @@ yum::Repository(
     """
     )
 
-    file_instances = project.get_instances(fortype="std::File")
-    assert len(file_instances) == 1
-    expected_content_repo_file = """
+    def get_expected_config_file() -> str:
+        result = """
 [test]
 name = test
 enabled=1
 gpgcheck=1
-baseurl = http://baseurl.com
-mirrorlist = http://mirror.com
+        """.strip()
+        if baseurl:
+            result = f"{result}\nbaseurl = http://baseurl.com"
+        if mirrorlist:
+            result = f"{result}\nmirrorlist = http://mirror.com"
+        if metalink:
+            result = f"{result}\nmetalink = http://metalink.com"
+        return result + "\n" + """
 gpgkey = http://gpgkey.com
 metadata_expire = 7200
 skip_if_unavailable=0
-    """.strip()
-    assert file_instances[0].content.strip() == expected_content_repo_file
+        """.strip()
+
+    file_instances = project.get_instances(fortype="std::File")
+    assert len(file_instances) == 1
+    assert file_instances[0].content.strip() == get_expected_config_file()
     assert file_instances[0].path == "/etc/yum.repos.d/test.repo"